switch to mysql
had to futz a coupel things. first, date were returned as timestamps, breaking gwt serialization.
Alos, looking up a long ID w/ and int breaks in an odd "no parameter for N"...

This checkin has some tobedeleted NestedCallback stuff going on.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@93 b8d5060c-ad15-0410-9d10-96fc93d4f2a4

+++ src/com/aavu/server/dao/UserDAO.java
@@ -12,7 +12,7 @@ public interface UserDAO {
 	ServerSideUser loadUserByUsername(final String username) throws UsernameNotFoundException, DataAccessException;
 	void save(User user);
 	List<User> getAllUsers();
-	User getUserForId(Integer id);
+	User getUserForId(long id);
 	void delete(User user);
 	User getUserByUsername(String username);
 }
